+++ src/main/java/mysql/modules/paddlesubscriptions/PaddleData.java
@@ -2,17 +2,17 @@ package mysql.modules.paddlesubscriptions;
 
 public class PaddleData {
 
-    private final int subId;
+    private final long subId;
     private final long userId;
     private final boolean unlocksServer;
 
-    public PaddleData(int subId, long userId, boolean unlocksServer) {
+    public PaddleData(long subId, long userId, boolean unlocksServer) {
         this.subId = subId;
         this.userId = userId;
         this.unlocksServer = unlocksServer;
     }
 
-    public int getSubId() {
+    public long getSubId() {
         return subId;
     }
